srv definition

+++ source/runtime/modules/render/include/PixelFormat.h
@@ -6,7 +6,7 @@
 #define PIXELFORMAT_H
 #include <cstdint>
 /* from VkImageFormat vk_core=1.3 */
-enum EPixelFormat: uint32_t {
+enum EPixelFormat: uint8_t {
     PF_UNDEFINED                                      = 0,
     PF_R4G4_UNORM_PACK8                               = 1,
     PF_R4G4B4A4_UNORM_PACK16                          = 2,
